Changes: 리눅스 환경 컴파일 추가           - Fedora KDE에서만 테스트되었습니다!           - 윈도우에서는 ./build 리눅스에서는 ./buildForLinux에 폴더가 생성됩니다.
