+++ .env.Unix.cmake
@@ -1,0 +1,2 @@
+# Optional set TINYOBJ_PATH to target specific version, otherwise defaults to external/tinyobjloader
+set(TINYOBJ_PATH /run/media/paperTTowel/System/Dev/2025/littleVulkanGameEngine/external/tinyobjloader)
