+++ src/lve_buffer.cpp
