+++ src/lve_buffer.hpp
