+++ src/lve_descriptors.cpp
