+++ src/lve_model.cpp
