+++ src/lve_renderer.cpp
