+++ src/lve_utils.hpp
