+++ src/systems/point_light_system.cpp
