+++ src/systems/simple_render_system.cpp
