+++ src/systems/simple_render_system.hpp
