+++ unixBuild.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+mkdir -p buildForLinux
+cd buildForLinux
+cmake -S ../ -B .
+make && make Shaders && ./LveEngine
+cd ..
